ItemMaterial altered, all materials added
ItemMaterial now inherits from the ScriptableObject class, as that will have both better performance in a larger scale application of the experiment, and also is more in line with the design structure I have chosen for Materials in general. Additionally, all materials have been added in Metals, Textiles, and Woods.