Metal-based Item Generation 0.1 complete
Item generation for Metals Materials has reached it's Minimal Viable Product, in that it is playable to a satisfactory state. The next steps will be to complete this step for Woods, Leathers, and Cloths, and then to further expand the editor so that items may be custom generated, to play around with it.